plotdz: pass single arrays to streamquad_trapz_f32
The z array doesn't have to be float32, because it is just used in
plotting, but the dist and a arrays are passed to libtopotoolbox and
must be single precision floats.

+++ src/topotoolbox/stream_object.py
@@ -443,8 +443,8 @@ class StreamObject():
         if ax is None:
             ax = plt.gca()
         z = self.ezgetnal(z)
-        dist = np.zeros_like(z)
-        a = np.ones_like(z)
+        dist = np.zeros_like(z, dtype=np.float32)
+        a = np.ones_like(z, dtype=np.float32)
 
         # Compute upstream distance using streamquad_trapz_f32
         # Another traversal might be more efficient in the future
